+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased](https://github.com/NWQMC/WQP_UI/compare/WQP_UI-5.22.0...master)
 ### Changed
 -   All javascript tests must now be imported into assets/test/js/karmaEntrySpec.js. A single test bundle is now created rather than a bundle for each spec.
+### Fixed
+-   When NLDI is queried, the distance parameter is not used if no distance is specified
 
 ## [5.22.0](https://github.com/NWQMC/WQP_UI/compare/WQP_UI-5.21.0...WQP_UI-5.22.0) - 2020-06-24
 ### Changed